finishes removing duplicate code in tenant admin module
