+++ webapp/src/main/webapp/src/app/admin/tenant/component/property-override-table-controls/property-override-table-controls.component.less
