+++ webapp/src/main/webapp/src/app/admin/tenant/component/property-override-table/property-override-table.component.less
