+++ webapp/src/main/webapp/src/app/admin/tenant/component/tenant-link/tenant-link.component.html
