+++ webapp/src/main/webapp/src/app/admin/tenant/component/toggle-label/toggle-label.component.less
