+++ webapp/src/main/webapp/src/app/admin/tenant/component/well-group-list/well-group-list.component.ts
