+++ webapp/src/main/webapp/src/app/admin/tenant/model/configuration-property.ts
